Commander refactor (#20)
* remove commander and list commands for cli import

* remove commander

* remove bin from package.json

* shorten descriptions
