+++ packages/slate-tools/src/commands/start.js
@@ -1,16 +1,18 @@
-const spawn = require('cross-spawn');
-const debug = require('debug')('slate-tools:start');
-const config = require('./includes/config');
+import spawn from 'cross-spawn';
+import debug from 'debug';
+import config from '../config';
 
-module.exports = function(program) {
+const logger = debug('slate-tools:start');
+
+export default function(program) {
   program
     .command('start')
     .alias('s')
     .description('Deploy theme and watch for file changes.')
     .option('-e, --environment [environment]', 'deploy to a comma-separated list of environments', 'development')
     .action((options = {}) => {
-      debug(`--gulpfile ${config.gulpFile}`);
-      debug(`--cwd ${config.themeRoot}`);
+      logger(`--gulpfile ${config.gulpFile}`);
+      logger(`--cwd ${config.themeRoot}`);
 
       const args = ['--gulpfile', config.gulpFile, '--cwd', config.themeRoot, '--environment', options.environment];
 
@@ -19,4 +21,4 @@ module.exports = function(program) {
         stdio: 'inherit',
       });
     });
-};
+}
